requirements.txt
pytest
pytest-cov
moto
mkdocs-material
mkdocstrings
python-dotenv
aws-lambda-powertools[pydantic]
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-dynamodb
aws-cdk.aws-apigatewayv2
aws-cdk.aws-apigatewayv2-integrations